+++ src/RcsbGroupWeb/RcsbGroupView/RcsbGroupChart/GroupChartAdditionalProperties.ts
@@ -1,5 +1,13 @@
 import {ChartMapType} from "./GroupChartLayout";
 
+
 export interface GroupChartAdditionalProperties {
     componentMountCallback(chartMap: ChartMapType, layout:string[]):void;
+    layoutConfig: LayoutConfigInterface;
+}
+
+export interface LayoutConfigInterface {
+     [layoutKey:string]:{
+            title:string|undefined;
+     }
 }