+++ src/RcsbGroupWeb/RcsbGroupView/RcsbGroupChart/RcsbGroupChartComponent.tsx
@@ -3,7 +3,7 @@ import {SearchQueryComponentFactory} from "../RcsbGroupSeacrhQuery/SearchQueryCo
 import {ChartMapType, GroupChartLayout} from "./GroupChartLayout";
 import {GroupProvenanceId} from "@rcsb/rcsb-api-tools/build/RcsbDw/Types/DwEnums";
 import {SearchQuery} from "@rcsb/rcsb-api-tools/build/RcsbSearch/Types/SearchQueryInterface";
-import {GroupChartAdditionalProperties} from "./GroupChartAdditionalProperties";
+import {GroupChartAdditionalProperties, LayoutConfigInterface} from "./GroupChartAdditionalProperties";
 import {SearchQueryTools as SQT} from "../../../RcsbSeacrh/SearchQueryTools";
 import {GroupChartMap as GDCM} from "./GroupChartTools";
 import classes from "../RcsbGroupMembers/Components/scss/bootstrap-group-display.module.scss";
@@ -45,10 +45,20 @@ export class RcsbGroupChartComponent extends React.Component <RcsbGroupChartInte
     private async updateState(): Promise<void> {
         const layout: string[] = this.props.facetLayoutGrid ?? SQT.getFacetStoreFromGroupProvenance(this.props.groupProvenanceId).facetLayoutGrid;
         const chartMap: ChartMapType = await GDCM.getChartMap(this.props.groupProvenanceId,this.props.groupId,this.props.searchQuery);
+        if(this.props.additionalProperties?.layoutConfig)
+            applyLayoutConfig(layout, chartMap, this.props.additionalProperties?.layoutConfig);
         this.setState({layout,chartMap},()=>{
             if(typeof this.props.additionalProperties?.componentMountCallback === "function")
                 this.props.additionalProperties.componentMountCallback(this.state.chartMap, this.state.layout);
         });
     }
 
+}
+
+function applyLayoutConfig(layout: string[], chartMap: ChartMapType, layoutConfig: LayoutConfigInterface){
+    layout.forEach(attr=>{
+        if(chartMap.has(attr) && layoutConfig[attr]){
+            chartMap.get(attr).chart.title = layoutConfig[attr].title;
+        }
+    });
 }